[Refactor] 도메인 모델 구조 개선 및 타입 통일 #256

+++ app/src/main/java/com/konkuk/medicarecall/data/api/elders/ElderRegisterService.kt
@@ -19,7 +19,7 @@ interface ElderRegisterService {
 
     @POST("elders/{elderId}/health-info")
     suspend fun postElderHealthInfo(
-        @Path("elderId") elderId: Int,
+        @Path("elderId") elderId: Long,
         @Body request: ElderHealthRegisterRequestDto,
     ): Response<Unit>
