+++ app/src/main/java/com/konkuk/medicarecall/data/repository/HealthRepository.kt
@@ -4,5 +4,5 @@ import com.konkuk.medicarecall.domain.model.Health
 import java.time.LocalDate
 
 interface HealthRepository {
-    suspend fun getHealth(elderId: Int, date: LocalDate): Result<Health>
+    suspend fun getHealth(elderId: Long, date: LocalDate): Result<Health>
 }
